requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
anyio==4.9.0
asgiref==3.8.1
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
disutils==1.4.32.post2
Django==5.2.1
django-scheduler==0.10.1
dnspython==2.7.0
fake-useragent==2.2.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icalendar==6.2.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
loguru==0.7.3
MarkupSafe==3.0.2
multidict==6.4.3
narwhals==1.38.2
numpy==2.2.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
propcache==0.3.1
protobuf==6.30.2
pyarrow==20.0.0
pydeck==0.9.1
pymongo==4.12.1
pyotp==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
selenium==4.32.0
setuptools==80.4.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
trio==0.30.0
trio-websocket==0.12.2
twscrape==0.17.0
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
base58==2.1.1
motor==3.3.2
extra-streamlit-components==0.1.71
solders==0.21.0
fastapi==0.104.1
uvicorn==0.24.0